requirements-server.txt
# Use with Python 3.12

PyPDF2==3.0.1
openai==1.85.0
requests==2.32.4
sseclient==0.0.27
python-dotenv==1.1.0
gunicorn==23.0.0
Flask==3.1.1
anki==25.2.7
nltk==3.9.1
numpy==2.1.0

sounddevice~=0.5.2
PyAudio~=0.2.14
overrides~=7.7.0
typeguard~=4.4.2
RapidFuzz~=3.13.0
future~=1.0.0
pandas~=2.2.3
soundfile~=0.13.1
scipy~=1.15.3
tensorflow~=2.19.0
pydantic~=2.11.4

huggingface-hub==0.32.5
llama-index-llms-huggingface==0.5.0
llama-index-llms-huggingface-api==0.5.0
llama-index==0.12.41
llama-index-storage-index-store-postgres==0.4.0
llama-index-vector-stores-postgres==0.5.3
llama-index-embeddings-huggingface==0.5.4
transformers[torch]==4.52.4
tf-keras==2.19.0
